Reorganize test utils
