+++ proving_system/plonk/src/test_utils/mod.rs
@@ -1,0 +1,8 @@
+/// A test circuit
+pub mod circuit_1;
+/// A test circuit
+pub mod circuit_2;
+/// Deserialize json to generate test circuits
+pub mod circuit_json;
+/// Useful tools to test plonk over different circuits
+pub mod utils;
