+++ proving_system/plonk/src/test_utils/utils.rs
@@ -1,0 +1,99 @@
+use lambdaworks_crypto::commitments::kzg::KateZaveruchaGoldberg;
+use lambdaworks_crypto::commitments::kzg::StructuredReferenceString;
+use lambdaworks_math::{
+    cyclic_group::IsGroup,
+    elliptic_curve::{
+        short_weierstrass::curves::bls12_381::{
+            curve::BLS12381Curve, field_extension::BLS12381PrimeField, twist::BLS12381TwistCurve,
+        },
+        traits::IsEllipticCurve,
+    },
+    field::{element::FieldElement, fields::montgomery_backed_prime_fields::U256PrimeField},
+    traits::IsRandomFieldElementGenerator,
+};
+use lambdaworks_math::{
+    elliptic_curve::short_weierstrass::curves::bls12_381::pairing::BLS12381AtePairing,
+    field::fields::montgomery_backed_prime_fields::IsModulus, unsigned_integer::element::U256,
+};
+
+/// Specific curves and fields are set for testing purposes
+pub const ORDER_R: U256 =
+    U256::from("73eda753299d7d483339d80809a1d80553bda402fffe5bfeffffffff00000001");
+
+#[derive(Clone, Debug)]
+pub struct FrConfig;
+impl IsModulus<U256> for FrConfig {
+    const MODULUS: U256 = ORDER_R;
+}
+
+pub type Curve = BLS12381Curve;
+pub type TwistedCurve = BLS12381TwistCurve;
+pub type FrField = U256PrimeField<FrConfig>;
+pub type FpField = BLS12381PrimeField;
+pub type FrElement = FieldElement<FrField>;
+pub type FpElement = FieldElement<FpField>;
+pub type Pairing = BLS12381AtePairing;
+pub type KZG = KateZaveruchaGoldberg<FrField, Pairing>;
+pub const ORDER_R_MINUS_1_ROOT_UNITY: FrElement = FrElement::from_hex("7");
+
+type G1Point = <BLS12381Curve as IsEllipticCurve>::PointRepresentation;
+type G2Point = <BLS12381TwistCurve as IsEllipticCurve>::PointRepresentation;
+
+/// Generates a test SRS for the BLS12381 curve
+/// n is the number of constraints in the system.
+pub fn test_srs(n: usize) -> StructuredReferenceString<G1Point, G2Point> {
+    let s = FrElement::from(2);
+    let g1 = <BLS12381Curve as IsEllipticCurve>::generator();
+    let g2 = <BLS12381TwistCurve as IsEllipticCurve>::generator();
+
+    let powers_main_group: Vec<G1Point> = (0..n + 3)
+        .map(|exp| g1.operate_with_self(s.pow(exp as u64).representative()))
+        .collect();
+    let powers_secondary_group = [g2.clone(), g2.operate_with_self(s.representative())];
+
+    StructuredReferenceString::new(&powers_main_group, &powers_secondary_group)
+}
+
+/// Generates a domain to interpolate: 1, omega, omega², ..., omega^size
+pub fn generate_domain(omega: &FrElement, size: usize) -> Vec<FrElement> {
+    (1..size).fold(vec![FieldElement::one()], |mut acc, _| {
+        acc.push(acc.last().unwrap() * omega);
+        acc
+    })
+}
+
+/// Generates the permutation coefficients for the copy constraints.
+/// polynomials S1, S2, S3.
+pub fn generate_permutation_coefficients(
+    omega: &FrElement,
+    n: usize,
+    permutation: &[usize],
+) -> Vec<FrElement> {
+    let identity = identity_permutation(omega, n);
+    let permuted: Vec<FrElement> = (0..n * 3)
+        .map(|i| identity[permutation[i]].clone())
+        .collect();
+    permuted
+}
+
+/// The identity permutation, auxiliary function to generate the copy constraints.
+fn identity_permutation(w: &FrElement, n: usize) -> Vec<FrElement> {
+    let u = ORDER_R_MINUS_1_ROOT_UNITY;
+    let mut result: Vec<FrElement> = vec![];
+    for index_column in 0..=2 {
+        for index_row in 0..n {
+            result.push(w.pow(index_row) * u.pow(index_column as u64));
+        }
+    }
+    result
+}
+
+/// A mock of a random number generator, to have deterministic tests.
+/// When set to zero, there is no zero knowledge applied, because it is used
+/// to get random numbers to blind polynomials.
+pub struct TestRandomFieldGenerator;
+impl IsRandomFieldElementGenerator<FrField> for TestRandomFieldGenerator {
+    fn generate(&self) -> FrElement {
+        FrElement::zero()
+    }
+}
